# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WAFFER",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://nad-api.squaretech.tech/api"
        },
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/admin/entities": {
            "get": {
                "tags": [
                    "Admin - Entities"
                ],
                "description": "Get all entities",
                "operationId": "get_all_entities",
                "parameters": [
                    {
                        "name": "with_paginate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "branch_type_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Entities"
                ],
                "description": "Add new entity.",
                "operationId": "8047119a449ee6d7c23144e28c5a4e94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name[ar]",
                                    "address",
                                    "email",
                                    "contact_number",
                                    "founding_date",
                                    "price_per_pet",
                                    "allowed_branches",
                                    "allowed_users",
                                    "branch_type_id"
                                ],
                                "properties": {
                                    "name[en]": {
                                        "type": "string"
                                    },
                                    "name[ar]": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "contact_number": {
                                        "type": "string"
                                    },
                                    "founding_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "national_id": {
                                        "type": "string"
                                    },
                                    "price_per_pet": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "allowed_branches": {
                                        "type": "integer"
                                    },
                                    "allowed_users": {
                                        "type": "integer"
                                    },
                                    "image": {
                                        "type": "file"
                                    },
                                    "branch_type_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/admin/entities/{id}": {
            "get": {
                "tags": [
                    "Admin - Entities"
                ],
                "description": "Get entity information.",
                "operationId": "show_entity",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Entities"
                ],
                "description": "Edit entity.",
                "operationId": "587f12e6aeb6b3001ef72dda945a2257",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name[en]": {
                                        "type": "string"
                                    },
                                    "name[ar]": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "contact_number": {
                                        "type": "string"
                                    },
                                    "founding_date": {
                                        "type": "string"
                                    },
                                    "price_per_pet": {
                                        "type": "string"
                                    },
                                    "allowed_branches": {
                                        "type": "string"
                                    },
                                    "allowed_users": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "file"
                                    },
                                    "branch_type_id": {
                                        "type": "integer"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Entities"
                ],
                "description": "Delete entered Entity.",
                "operationId": "delete_entity",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}